requirements.txt
line-bot-sdk>=2.4.2
flask>=2.0.1
python-dotenv>=0.19.0
requests>=2.31.0
openai>=1.0.0
dropbox>=11.36.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
googlemaps>=4.10.0
schedule>=1.2.0
python-dateutil>=2.8.2
feedparser>=6.0.10
apscheduler>=3.10.1
